Refactored, I was going to do more but we need to start doing work
