+++ views/wip.ejs
@@ -1,0 +1,4 @@
+<%- include('partials/header') %>
+  <h1>Work in Progress</h1>
+  <p>Come back later.</p>
+<%- include('partials/footer') %>
